requirements.txt
beautifulsoup4==4.12.2
openai==0.27.2
Pillow==9.5.0
requests==2.28.2
streamlit==1.18.0
textblob==0.17.1
torch==2.0.0
transformers==4.28.1
snowflake-connector-python==3.0.2
snowflake-snowpark-python==1.2.0